Replace 'L' with 'l' for species names.
The database uses upper-case L to distinguish from 1. With an
appropriate font this shouldn't be a problem and searching for "Cl" or
"Al" is more intuitive than "CL" or "AL". This also applies to the
bracketed suffixes denoting non-gaseous states/phases. This notation
has been reverted to lower-case l.

Obvious exceptions to this are elements denoted by Lx. In practice
only Lithium is going to be present here (it is, I've checked) but I
excluded the patterns La, Li, Lr, Lu and Lv from the transformation
just to be doubly sure.