Unit tests for the thermoinp module written.
Accompanying the unit tests is a set of changes to the module in order
to satisfy compliance.

  - An 'exact' flag has been added to thermoinp.lookup() for
    whole-name matching.
  - An 'exact' flag has been added to thermoinp.create_subset() for
    whole-name matching.
  - create_subset() output now matches the output of ThermoBuild. This
    is slightly different from the source data file but the assumption
	is that ThermoBuild output is correct. This doesn't actually
	affect the species datasets, only the skeleton file structure.
  - create_subset() now accepts a list/tuple or string of search
    strings for creating non-trivial databases.
  - The reference temperature for species without temperature
    intervals was not being coerced into a float. Fixed.

A bunch of indirect changes have also been made, including additional
comments, variable re-naming and slight (background) changes to some
data structures. The intent is to improve clarity and semantics, with
a very minor amount of refactoring/changing behaviour.