Shift to the NASAPoly interval class by default
  - Defaults within the Species class remain tied to 'Interval' for
    now, but otherwise the database instantiates with NASAPoly*
    intervals.
  - Choice of NASAPoly is provided as an optional argument for the
    database:

        '' : NASAPoly
        'nd' : NASAPolyND (methods to calculate non-dim properties)
        'ml' : NASAPolyML (methods to calculate molar properties)

Interval will be deprecated properly in future, the NASAPoly is more
flexible and allows for choice of functionality. The basic NASAPoly
actually provides everything required for constructing functions to
calculate properties (though documentation of this is lacking).
However, for convenience functions are provided in the ND and ML
variants (specific properties can also be calculated but not without
access to molwt, which is not stored within the poly spec).